+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Upgraded python-lib to v2.9.4
 
 ### Changed
+- run_program: Set flavor_id for ProgramLock to "elasticrepo".
 - config/elastic.py.TEMPLATE: Set new syntax for host entry.
 - requirements.txt: Added certifi==2019.11.28 and updated requests==2.6.0 entries.
 - Converted imports to use Python 2.7 or Python 3.